review: refactor to abstract URI certificate verification
