+++ pkg/config/dynamic/http_config.go
@@ -232,10 +232,7 @@ type ServersTransport struct {
 	MaxIdleConnsPerHost int                        `description:"If non-zero, controls the maximum idle (keep-alive) to keep per-host. If zero, DefaultMaxIdleConnsPerHost is used" json:"maxIdleConnsPerHost,omitempty" toml:"maxIdleConnsPerHost,omitempty" yaml:"maxIdleConnsPerHost,omitempty" export:"true"`
 	ForwardingTimeouts  *ForwardingTimeouts        `description:"Timeouts for requests forwarded to the backend servers." json:"forwardingTimeouts,omitempty" toml:"forwardingTimeouts,omitempty" yaml:"forwardingTimeouts,omitempty" export:"true"`
 	DisableHTTP2        bool                       `description:"Disable HTTP/2 for connections with backend servers." json:"disableHTTP2,omitempty" toml:"disableHTTP2,omitempty" yaml:"disableHTTP2,omitempty" export:"true"`
-	// CertVerifier is only (optionally) set by the Consul Catalog provider
-	// in order to implement the Consul Connect protocol between Traefik and the Consul services.
-	// Therefore it is not serializable to prevent it from being configurable by other parties.
-	CertVerifier CertVerifier `json:"-" toml:"-" yaml:"-" hash:"-" label:"-" export:"false"`
+	PeerCertURI         string                     `description:"URI used to match against SAN URI during the peer certificate verification." json:"peerCertURI,omitempty" toml:"peerCertURI,omitempty" yaml:"peerCertURI,omitempty" export:"true"`
 }
 
 // +k8s:deepcopy-gen=true
