+++ pkg/provider/consulcatalog/config_test.go
@@ -400,11 +400,7 @@ func Test_buildConfiguration(t *testing.T) {
 									KeyFile:  "key",
 								},
 							},
-							CertVerifier: newVerifierData(itemData{
-								Datacenter: "dc1",
-								Name:       "dev/Test",
-								Namespace:  "ns",
-							}),
+							PeerCertURI: "spiffe:///ns/ns/dc/dc1/svc/dev/Test",
 						},
 					},
 				},
@@ -490,11 +486,7 @@ func Test_buildConfiguration(t *testing.T) {
 									KeyFile:  "key",
 								},
 							},
-							CertVerifier: newVerifierData(itemData{
-								Datacenter: "dc1",
-								Name:       "dev/Test",
-								Namespace:  "ns",
-							}),
+							PeerCertURI: "spiffe:///ns/ns/dc/dc1/svc/dev/Test",
 						},
 					},
 				},
